% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_summary_metrics.R
\name{aggregate.subjects.wrong.type}
\alias{aggregate.subjects.wrong.type}
\title{Find, for each subject, the number of variables for
which that subject has data of the wrong type}
\usage{
\method{aggregate}{subjects.wrong.type}(variable.summary)
}
\arguments{
\item{variable.summary}{List; variable configuration and
summary data containing precomputed information about subjects
with unexpected type data for each variable.}
}
\value{
Named integer vector, where values are counts of non-compliant
variables per subject, and names are subject IDs.
}
\description{
This aggregates non-compliant subject ID information from
all variables and reduces them to counts of instances
per subject, for downstream tabular reporting.
}
\details{
These counts may be inflated per subject if there is
something specifically wrong with a block of related
variables. This count summary is hopefully a proxy
for subject-specific data corruption, so they can
be removed before downstream analysis.
}
\examples{
var.sum <- list(variables = list(
  HW00001 = list(subjects.wrong.type = c("A")),
  HW00002 = list(subjects.wrong.type = c("A", "B"))
))
all.bad.subjects <- process.phenotypes:::aggregate.subjects.wrong.type(var.sum)
}
\seealso{
apply.type.conversions
}
