% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_phenotype_cleaning.R
\name{apply.type.conversions}
\alias{apply.type.conversions}
\title{Apply type constraints to each variable from yaml config}
\usage{
apply.type.conversions(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{data frame, loaded phenotype data with
standardized headers; all columns should be character vectors}

\item{variable.summary}{list, per-column summary information
and parameters from yaml input}
}
\value{
list, entry 'phenotype.data' is a data frame (modified
version of phenotype input data with conversions applied as appropriate);
entry 'variable.summary' is a list (modified version of input list
with summary information injected from certain handlers)
}
\description{
A short list of special handlers are implemented for particular
data conversion problems (e.g. numerics with trailing unit suffixes,
or blood pressure data)
}
\details{
Each yaml configuration block should have a "type" entry
(if not, will default to unmodified string) denoting what
data storage type is expected for this variable.
}
\seealso{
load.configuration
}
\keyword{phenotypes}
\keyword{yaml}
