% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{collapse.repeats}
\alias{collapse.repeats}
\title{Basic global cleanup of entries in a phenotype data frame}
\usage{
collapse.repeats(
  df,
  targets = c("\\\\\\\\/", "\\\\."),
  replacements = c("/", ".")
)
}
\arguments{
\item{df}{data frame, input phenotype content}

\item{targets}{character vector, input character(s) to replace duplicates of}

\item{replacements}{character vector, character to replace target duplicates with}
}
\value{
modified version of input with values cleaned as described
}
\description{
Collapses multiple consecutive instances of specificed character;
defaults to operating on slashes and periods
}
\details{
Performs the following data modifications to non-numeric values:
\itemize{
\item collapses multiple instances of specified character to a single replacement
}
}
