% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_lists.R
\name{combine.lists}
\alias{combine.lists}
\title{Combine lists (or other named structure)
by overwriting specific entries in substructures}
\usage{
combine.lists(list1, list2)
}
\arguments{
\item{list1}{list or vector, existing list/vector that will have
entries added into it}

\item{list2}{list or vector, derived list/vector that will have
its entries added to the other}
}
\value{
list or vector, combined version of input objects
with list1 default entries present unless they
are overridden by identically named entries
in list2, and any unique entries in list2
added on top.
}
\description{
Combine lists (or other named structure)
by overwriting specific entries in substructures
}
\details{
Configuration lists from yaml format should ideally
be merged by overwriting specific entries if present,
while leaving any other existing entries intact. This
doesn't appear to be available by default in R, so
this utility function does the required DFS to
enable this functionality.
}
