% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_files.R
\name{compare.column.pair}
\alias{compare.column.pair}
\title{Compare a column to another column for full tabular report}
\usage{
compare.column.pair(col1, col1.name, col1.config, col2, col2.name, col2.config)
}
\arguments{
\item{col1}{vector, first data vector to be compared}

\item{col1.name}{character vector, name of first data vector
in its dataset}

\item{col1.config}{list, configuration data for first vector}

\item{col2}{vector, second data vector to be compared}

\item{col2.name}{character vector, name of second data vector}

\item{col2.config}{list, configuration data for second vector
in its dataset}
}
\description{
TBD
}
\details{
this function is designed to be used if two approximately
(or ideally) identical columns have been found, and the overlap
between their results needs to be determined
}
