% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_files.R
\name{compare.columns}
\alias{compare.columns}
\title{Compare a column to another data frame for approximate identity}
\usage{
compare.columns(query.column, query.colname, target.df)
}
\arguments{
\item{query.column}{vector, data to be queried against target columns}

\item{query.colname}{character vector, name of query column
in its dataset}

\item{target.df}{data frame, target dataset to scan against query
column}
}
\description{
TBD
}
\details{
a single column is compared against all columns of a data frame,
to try to find the most likely identical column in the target.
this is likely to completely fail until at least minimal
initial processing with the main methods of this
package.
}
