% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_files.R
\name{compare.files}
\alias{compare.files}
\title{Compare two files for approximate identity of columns}
\usage{
compare.files(
  first.filename,
  first.config.yaml,
  second.filename,
  second.config.yaml,
  second.dataset.tag,
  shared.model.yaml
)
}
\arguments{
\item{first.filename}{character vector, name of
first file to load and compare. should be the
file with identity tags linking its variables
to the second dataset}

\item{first.config.yaml}{character vector, name
of first file's config yaml}

\item{second.filename}{character vector, name of
second file to load and compare}

\item{second.config.yaml}{character vector, name
of second file's config yaml}

\item{second.dataset.tag}{character vector,
dataset tag for second input; used to find
TAG##### variable identifiers in yaml
identity entries}

\item{shared.model.yaml}{character vector,
name of shared model configuration for both
input files

note that this function is primary used for its
side effect, emitting content to screen.}
}
\description{
TBD
}
\details{
as a utility function, this compares two datasets
for approximately identical columns, based not on
the header labels but on column values. this is
likely to completely fail until at least minimal
initial processing with the main methods of this
package.
}
