% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_files.R
\name{compare.files.report}
\alias{compare.files.report}
\title{Generate an html report of pre- and post-audit
comparison}
\usage{
compare.files.report(
  first.filename,
  first.config.yaml,
  second.filename,
  second.config.yaml,
  shared.model.yaml,
  out.filename
)
}
\arguments{
\item{first.filename}{character vector, name of
first file to load and compare. should be the
file with identity tags linking its variables
to the second dataset}

\item{first.config.yaml}{character vector, name
of first file's config yaml}

\item{second.filename}{character vector, name of
second file to load and compare}

\item{second.config.yaml}{character vector, name
of second file's config yaml}

\item{shared.model.yaml}{character vector,
name of shared model configuration for both
input files

note that this function is primary used for its
side effect, emitting content to screen.}
}
\description{
TBD
}
\details{
Calling rmarkdown directly is hard.
}
