% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_summary_metrics.R
\name{compute.subject.na.count}
\alias{compute.subject.na.count}
\title{Compute per-subject NA count across all variables.}
\usage{
compute.subject.na.count(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{Data frame of subject data, with subjects
as rows and phenotypes as columns.}

\item{variable.summary}{List; variable configuration and
summary data with subject ID variable annotated as \code{subject_id = TRUE}.}
}
\value{
Named integer vector, where values are counts of NAs
for each subject across all variables, and names are subject IDs.
}
\description{
Computes a count for each subject of all NAs across
all variables, after all filtering and conversion has
been applied.
}
\details{
NAs can be present in input data or introduced at various
steps, and the overall count can be indicative of various
phenomena, including expected blockwise NA due to individual
subjects not completing all questionnaire sections.
}
\examples{
pheno.data <- data.frame(
  HW00001 = c("A", "B", "C", "D"),
  HW00002 = c(1, 2, NA, 3)
)
var.sum <- list(variables = list(
  HW00001 = list(
    name = "name1",
    type = "string",
    subject_id = TRUE
  ),
  HW00002 = list(
    name = "name2",
    type = "string"
  )
))
subject.na.counts <- compute.subject.na.count(pheno.data, var.sum)
}
\seealso{
apply.type.conversions
}
