% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{config.validation}
\alias{config.validation}
\title{Run schema validation for process.phenotypes configuration files}
\usage{
config.validation(
  dataset.yaml,
  shared.models.yaml,
  dataset.schema,
  shared.models.schema
)
}
\arguments{
\item{dataset.yaml}{character; name of input dataset yaml file}

\item{shared.models.yaml}{character; name of input shared models yaml file}

\item{dataset.schema}{character; name of schema specification for dataset yaml}

\item{shared.models.schema}{character; name of schema specification for shared models yaml}
}
\description{
The function returns NULL, and is run primarily for its side effect
of emitting verbose logging data about the manner of validation
error detected by the json_validate call.
}
