% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_phenotype_cleaning.R
\name{convert.type}
\alias{convert.type}
\title{Coordinate different function calls for type conversion}
\arguments{
\item{vec}{character vector, input phenotype content}

\item{var.summary}{list, variable summary entry for this particular variable}

\item{target.type}{character vector, denotes the expected type for each variable;
read from the yaml config}
}
\value{
list, 'phenotype.data' contains converted phenotype information,
'variable.summary' contains input variable summary.
}
\description{
Each yaml configuration block should have a "type" entry
denoting what data storage type is expected for this variable.
This function dispatches the appropriate handlers for these
methods, specifically triggering type-specific cleaning in some
cases before casting the values.

If for any reason you don't like the type conversion behaviors
in this package, set the variable type to "string" and then
implement the processing you desire as a derived variable.
}
\details{
This function is a pure internal, and should never be
called outside of the package.
}
\seealso{
apply.type.conversions
}
\keyword{phenotypes}
\keyword{yaml}
