% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{create.derived.variable}
\alias{create.derived.variable}
\title{Apply config-specified R commands to generate derived variables}
\usage{
create.derived.variable(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{data.frame, input phenotype data}

\item{variable.summary}{list, configuration data per variable}
}
\value{
list, modified version of input variable.summary argument
with some reporting information about dependency check success/failure
}
\description{
TBD
}
\details{

}
