% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phenotype_report.R
\name{create.phenotype.report}
\alias{create.phenotype.report}
\title{Create markdown report summarizing phenotype data in file}
\usage{
create.phenotype.report(
  in.filename,
  dataset.tag,
  dataset.yaml,
  shared.model.yaml,
  out.filename,
  magic.fix = TRUE,
  quote = ""
)
}
\arguments{
\item{in.filename}{character vector, name of input phenotype tsv file}

\item{dataset.tag}{character vector, dataset-specific tag to prefix variable names}

\item{dataset.yaml}{character vector, yaml configuration for project}

\item{shared.model.yaml}{character vector, yaml configuration for shared model specifications}

\item{out.filename}{character vector, name of output report html file}

\item{quote}{character vector, character used to quote string tokens; defaults to null}
}
\description{
TBD
}
\details{
Given phenotype information in tsv file, run some TBD reporting/cleaning
and emit summary information.
}
\examples{
create.phenotype.report("/path/to/directory/MM_FINAl_store.tsv", "MM", "/output/path/MM_report.html")
}
\seealso{
run.experiment
}
\keyword{phenotypes}
