% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{detect.unicode.characters}
\alias{detect.unicode.characters}
\title{Detects and reports residual Unicode characters}
\usage{
detect.unicode.characters(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{data.frame, input phenotype data}

\item{variable.summary}{list, per-variable summary information and config data}
}
\value{
list, input variable summary with updated logging information
if any residual Unicode characters are detected.
}
\description{
TBD
}
\details{
Unicode characters should be removed when detected, as they are typically
only sporadically injected by upstream text processors and will interfere
with any downstream attempts at standardization. \code{process.unicode.characters}
attempts to replace an existing set of these nefarious characters with their
ASCII-style equivalents, but others may be introduced later, and can
be flagged here.
}
