% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_phenotype_cleaning.R
\name{exclude.by.missing.subject.id}
\alias{exclude.by.missing.subject.id}
\title{Exclude subjects for whom no ID has been recorded}
\usage{
exclude.by.missing.subject.id(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{Data frame of loaded phenotype data with
standardized headers. All columns should be character vectors.}

\item{variable.summary}{List of per-column summary information
and parameters from dataset configuration.}
}
\value{
List, with first entry 'phenotype.data' data frame of input
phenotypes with subjects without a subject ID excluded; and second
entry 'variable.summary' input configuration with information about
excluded subjects.
}
\description{
This function excludes subjects where their ID is NA, as
such subjects cannot be meaningfully linked to relevant external
data sources, most notably consent calls.
}
\details{
This function was added during the original set of dataset processing
that led to the development of this package. More recent input datasets
have almost completely lacked this particular discrepancy. If any
subjects whatsoever are flagged as removed by this function, something
is likely seriously malformed in the input data.
}
\examples{
phenotype.data <- data.frame(
  HW00001 = c("A", "B", NA, "D"),
  HW00002 = 1:4
)
variable.summary <- list(variables = list(
  HW00001 = list(params = list(
    name = "subjid",
    type = "string",
    subject_id = TRUE
  )),
  HW00002 = list(params = list(
    name = "count",
    type = "numeric"
  ))
))
result <- process.phenotypes:::exclude.by.missing.subject.id(phenotype.data, variable.summary)
}
