% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject_exclusions.R
\name{exclude.subjects.by.metric}
\alias{exclude.subjects.by.metric}
\title{Remove subjects from a data frame based on a pre-computed
metric and a bound on that metric.}
\usage{
exclude.subjects.by.metric(phenotype.data, metric.vec, upper.limit)
}
\arguments{
\item{phenotype.data}{data.frame, input phenotype data}

\item{metric.vec}{numeric vector, pre-computed values,
one per row of input data frame, to be used for determining
exclusion}

\item{upper.limit}{numeric, bound on computed metric
above which corresponding subjects will be removed}
}
\description{
Assuming metrics like per-subject type conversion failure
counts have been computed elsewhere, this function removes
rows exceeding some bound on that value.
}
