% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_dependencies.R
\name{find.subject.id.index}
\alias{find.subject.id.index}
\title{Locate a variable in the dataset flagged as subject ID}
\usage{
find.subject.id.index(variable.summary)
}
\arguments{
\item{variable.summary}{List of input configuration data.}
}
\value{
Integer index of subject ID column in input data.
}
\description{
The input yaml specification allows a single variable
to be flagged as "subject_id: true" for context-specific
behaviors. This function scans the configuration data and
returns the location of that variable.
}
\details{
If this column is not located, stop() is called. The handling
of this tag, and the \code{subject_age} tag, could be improved.
}
\examples{
variable.summary <- list(variables = list(
  HW00001 = list(params = list(
    name = "subjid",
    type = "string",
    subject_id = TRUE
  )),
  HW00002 = list(params = list(
    name = "subjage",
    type = "string"
  ))
))
subj.id.index <- process.phenotypes:::find.subject.id.index(variable.summary)
}
