% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{flag.required.variables}
\alias{flag.required.variables}
\title{Add flags for subject ID and age variables}
\usage{
flag.required.variables(out.yaml, subject.id.name, age.name)
}
\arguments{
\item{out.yaml}{list; constructed variable configuration data}

\item{subject.id.name}{character vector; expected name of subject ID variable}

\item{age.name}{character vector; expected name of age variable}
}
\value{
list; input yaml with flags added to appropriate variables
}
\description{
Add flags for subject ID and age variables
}
