% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{generate.predicted.yaml}
\alias{generate.predicted.yaml}
\title{For repeat expansion: run an internal pass
of parse.surveycto to reuse the logic for
generating configuration blocks in case new
blocks have magically appeared in new SurveyCTO
exports}
\usage{
generate.predicted.yaml(
  existing.yaml,
  form.definition.filename,
  new.data.filename,
  intermediate.dataset.yaml,
  intermediate.shared.models
)
}
\arguments{
\item{existing.yaml}{list; existing dataset configuration
loaded with read_yaml. Note that the entire configuration
need not be correctly specified, but the variables containing
the mandatory 'subject_id' and 'subject_age' tags should
be correctly specified and ordered}

\item{form.definition.filename}{character; name of
SurveyCTO xlsx form definition}

\item{new.data.filename}{character; name of wide csv
export from SurveyCTO}

\item{intermediate.dataset.yaml}{character; name of
output dataset yaml from parse.surveycto}

\item{intermediate.shared.models}{character; name
of output shared models yaml from parse.surveycto}
}
\description{
For repeat expansion: run an internal pass
of parse.surveycto to reuse the logic for
generating configuration blocks in case new
blocks have magically appeared in new SurveyCTO
exports
}
