% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{get.bins}
\alias{get.bins}
\title{Make sure the number of bins in a histogram isn't too close to the
number of unique values!}
\usage{
get.bins(vec)
}
\arguments{
\item{vec}{numeric vector; input data to histogram}
}
\description{
Make sure the number of bins in a histogram isn't too close to the
number of unique values!
}
