% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{get.last.variable.number}
\alias{get.last.variable.number}
\title{Determine from a set of variables what the last
variable number was and return as integer}
\usage{
get.last.variable.number(variables)
}
\arguments{
\item{variables}{list of previously added variables,
with names of the format "TAG#####.*"}
}
\value{
numeric part of final named entry as integer
}
\description{
Determine from a set of variables what the last
variable number was and return as integer
}
