% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{get.top.ten}
\alias{get.top.ten}
\title{Helper function to create highest/lowest ten tables in markdown report.}
\usage{
get.top.ten(decreasing, vec, column.label)
}
\arguments{
\item{decreasing}{logical, passed to sort to control increasing or decreasing order}

\item{vec}{named vector, containing values and counts}

\item{column.label}{string, name of first column (e.g. Subjects, Variables)}
}
\value{
data.frame with ten rows containing values and counts
}
\description{
Takes a named vector containing values (e.g. subjects) and counts,
sorts by number of counts, takes the top ten, and returns a data.frame
}
\details{
This is a helper function used to grab the highest or lowest ten values
to print as a table in the markdown report.
}
