% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{handle.existing.block}
\alias{handle.existing.block}
\title{For repeat expansion, handle the situation where there
is a repeat variable that was present
in the original configuration file, but that has a higher
number repeat present in the current SurveyCTO export}
\usage{
handle.existing.block(var.prefix, trailing.num, res.variables, missing.colname)
}
\arguments{
\item{var.prefix}{character; shared common prefix of the dataset
yaml configuration for this repeat variable. So for example,
if the existing repeat block was "HW00001_1", "HW00001_2", "HW00001_3",
this would be "HW00001_"}

\item{trailing.num}{numeric; instance of the repeat observed
in the new dataset but absent from original configuration}

\item{res.variables}{list; accumulator of output variable
configuration with repeats added or extended as needed}

\item{missing.colname}{character; encountered column name
that is missing from current configuration. Expected to be
correctly predicted and present in parse.surveycto configuration
based on the current form definition and csv export}
}
\value{
list; updated version of input parameter 'res.variables'
with the new resolved variable configuration entry added
}
\description{
For repeat expansion, handle the situation where there
is a repeat variable that was present
in the original configuration file, but that has a higher
number repeat present in the current SurveyCTO export
}
