% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{handle.missing.block}
\alias{handle.missing.block}
\title{For repeat expansion, handle the situation where there
is a repeat variable that was not present in any form
in the original configuration file.}
\usage{
handle.missing.block(
  varname,
  regenerated.yaml,
  max.existing.number,
  res.variables,
  missing.colname
)
}
\arguments{
\item{varname}{string; formatted variable name for predicted
insertion point}

\item{regenerated.yaml}{list; loaded dataset yaml corresponding
to current actual wide csv export from SurveyCTO. Expected to
be the loaded output of parse.surveycto}

\item{max.existing.number}{numeric; maximum variable number,
expecting variable names of the format "HW#####"}

\item{res.variables}{list; accumulator of output variable
configuration with repeats added or extended as needed}

\item{missing.colname}{character; encountered column name
that is missing from current configuration. Expected to be
correctly predicted and present in parse.surveycto configuration
based on the current form definition and csv export}
}
\value{
list; res.variables, updated version of input parameter
with the new resolved variable configuration entry added; max.existing.number,
updated version of input parameter incremented to reflect variable addition
}
\description{
For repeat expansion, handle the situation where there
is a repeat variable that was not present in any form
in the original configuration file.
}
