% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ancestry_labels.R
\name{harmonize.ancestry.from.linker}
\alias{harmonize.ancestry.from.linker}
\title{Update ancestry variable with harmonized ancestry names,
and report how well you did}
\usage{
harmonize.ancestry.from.linker(
  phenotype,
  variable,
  ancestry.data,
  best.match.threshold = 0.9,
  best.match.discernment = 0.75
)
}
\arguments{
\item{phenotype}{character vector, input self-reported ancestry}

\item{variable}{list, configuration data for input phenotype}

\item{ancestry.data}{named character vector, linking from unique
possible input ancestries to harmonized ancestry label. note that
both entries and names in this vector are assumed to have gone
through approximately analogous processing as the input free-text
ancestry group labels, e.g. made lowercase and had whitespace removed.}

\item{best.match.threshold}{numeric, on [0,1], denoting how high
the best match similarity must be to permit a call}

\item{best.match.discernment}{numeric, on [0,1] denoting the proportion
of the top match the next best match must be less than (assuming the
next best match does not map to the same harmonized label) in order
to accept the top match}
}
\value{
list, first entry is harmonized phenotype data as a factor
with levels set to the available ancestry levels in the known ancestry
input, second entry is input configuration list with added metrics
about mapping performance
}
\description{
This function attempts to map free-text ancestry data
to a set of fixed ancestry labels based on an input linker
between known ancestries and harmonized labels. Exact matches
are assigned as expected. Then, an attempt is made to harmonize
partial matches based on closest match. Actual algorithm TBD.
}
\details{
Very much TBD
}
\seealso{
load.ancestry.linker
}
