% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{is.blood.pressure}
\alias{is.blood.pressure}
\title{Detect whether strings look like SBP/DBP measurements}
\usage{
is.blood.pressure(vec, allow.trailing = FALSE)
}
\arguments{
\item{vec}{character vector, input candidate values}

\item{allow.trailing}{logical, whether a blood pressure-like
entry with arbitrary trailing text should be considered possibly
a good result}
}
\value{
logical vector, one per input value, whether the input
matches blood pressure format
}
\description{
Attempts to identify blood pressure measurements in
DBP/SBP format
}
\details{
Blood pressure measurements may be reported as ###/###
entries in a single vector, complicating numeric casts.
}
