% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ancestry_labels.R
\name{load.ancestry.linker}
\alias{load.ancestry.linker}
\title{Read ancestry linker data and format it for use in R}
\usage{
load.ancestry.linker(filename)
}
\arguments{
\item{filename}{character vector, name of file containing
structured linker data}
}
\value{
character vector, named vector of ancestries;
names are unique expected inputs in phenotype dataset,
values are harmonized labels to be applied in the output
}
\description{
Free-text self-reported ancestry is expected to be heterogeneous
and complicated. For use in genomics contexts, the labels should
be harmonized, at the very least applying single standardized
group labels to groups with multiple possible (typically quite
similar) names.
}
\details{
Input ancestry linker format is based off the source of
ancestry information from the phase 2 questionnaire, which
appears to actually be
https://en.wikipedia.org/wiki/List_of_ethnic_groups_in_Nigeria
(accessed 17 August 2021). So, in short, the first and only
column (tab-delimited) should be formatted as
"Name (Alt1, Alt2, ...)",
with one row per unique possible group in the output. No
header should be present. Note that triple colons are not
allowed in any ancestry group name for the time being.
}
