% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{make.lowercase}
\alias{make.lowercase}
\title{Harmonize all input phenotype entries to lowercase}
\usage{
make.lowercase(df)
}
\arguments{
\item{df}{Data frame of input phenotype data as character vectors.}
}
\value{
Data frame containing input entries with all alphabetical
values set to their lowercase versions.
}
\description{
To facilitate downstream operations comparing values
to configuration data and each other, all string input
values are set to lowercase. This is primarily to
address issues with files that have been constructed
via manual entry.
}
\details{
The output data frame will have all string values in lowercase.
This is usually fine, but can cause some issues when input
subject IDs contain uppercase letters. This package assumes
that subject IDs are not ultimately case sensitive. If this causes
substantial problems for users, we could potentially patch this
to restore ID case in the output data; post an issue if you are
interested in this functionality.
}
\examples{
phenotype.data <- data.frame(
  HW00001 = c("A", "B", "C", "D"),
  HW00002 = c("1.03mL", "2.33", "2.22 L", "4")
)
result <- process.phenotypes:::make.lowercase(phenotype.data)
}
