% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_phenotype_cleaning.R
\name{parse.date}
\alias{parse.date}
\title{Convert values that should be dates to year}
\usage{
parse.date(vec, var.summary)
}
\arguments{
\item{vec}{character vector, input phenotype content}

\item{var.summary}{list, variable summary entry for this particular variable}
}
\value{
modified version of input with values cleaned as described above
}
\description{
Extracts two or four digit year values from dates of several formats,
assigns likely century (20 or 19), and replaces the original date
with just the four-digit year value.  Also removes extremely low
likely erroneous year values.
}
\details{
Given string vector with malformed date entries, attempt
to force conversion to YYYY
}
