% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{print.conditionally}
\alias{print.conditionally}
\title{Helper function to conditionally print output content}
\usage{
\method{print}{conditionally}(val)
}
\arguments{
\item{val}{thing to be conditionally printed}
}
\description{
If the passed value is null, no action is taken; otherwise,
the thing is emitted wrapped in a print() or cat() statement
}
