% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{reformat.blood.pressure}
\alias{reformat.blood.pressure}
\title{Detect and reformat blood pressure measures}
\usage{
reformat.blood.pressure(vec, var.summary)
}
\arguments{
\item{vec}{character vector, input phenotype content}

\item{var.summary}{list, variable summary information for this particular variable}
}
\value{
modified version of input with values cleaned as described above
}
\description{
Removed unexpected values from probable blood pressure entries
}
\details{
Given string vector with potentially malformed blood pressure
entries, attempt to standardize to SBP/DBP by stripping
common problems.
}
