% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{reformat.factor}
\alias{reformat.factor}
\title{Convert character vector to factor, respecting configuration aliases}
\usage{
reformat.factor(vec, variable.summary)
}
\arguments{
\item{vec}{Character vector of input phenotype data.}

\item{variable.summary}{List of input configuration data for the input variable.}
}
\value{
List, with first entry 'phenotype.data' containing factor converted
phenotype information, and second entry 'variable.summary' containing input
configuration with possible additional information
about invalid factor levels converted to NA.
}
\description{
Categorical, ordinal, and binary type configured variables in the input
dataset configuration are converted into factors. Factors are matched
against each level's primary name. In situations
where harmonization is required, for example to create compatible factor
levels with other variables or rename factor levels, the \code{alternate_patterns}
sequence of regular expressions is used to detect and alias responses to harmonized
factor levels.
}
\details{
Note that pattern matching is expected to be conducted versus data inputs
that have already passed through upstream cleaning and have been made lowercase,
had nonword characters removed, etc. If there are issues, these will be reported
as invalid type conversions in the \code{create.phenotype.report} cleaning report,
and the specified patterns can be updated accordingly.
}
\examples{
data <- c("apple", "banana", "red apple")
variable.summary <- list(params = list(
  name = "fruit",
  type = "categorical",
  levels = list(
    "1" = list(
      name = "apple",
      alternate_patterns = c("red apple")
    ),
    "2" = list(name = "banana")
  )
))
results <- process.phenotypes:::reformat.factor(data, variable.summary)
}
