% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{reformat.numerics}
\alias{reformat.numerics}
\title{Aggressively convert malformed numeric vectors}
\usage{
reformat.numerics(df, accept.proportion = 0.75)
}
\arguments{
\item{df}{data frame, input phenotype content}

\item{accept.proportion}{numeric, proportion of data to match a type
required to enforce the match}
}
\value{
modified version of input with values cleaned as described above
}
\description{

}
\details{
Given string vectors with malformed numeric entries, attempt
to force conversion to numerics by stripping common problems.
}
\examples{
phenotype.data <- data.frame(
  A = c("yes", "yes", "no", NA),
  B = c("1.0", "100/80", "4g", "sparse wrong value")
)
phenotype.data <- reformat.numerics(phenotype.data)
}
