% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.bp.ratio}
\alias{report.bp.ratio}
\title{Helper function to report plot showing relationship between
self-reported systolic and diastolic blood pressure}
\usage{
report.bp.ratio(
  phenotype.data,
  variable.entry,
  name,
  my.theme,
  suppress.reporting
)
}
\arguments{
\item{phenotype.data}{data.frame; full phenotype data for variable
selection}

\item{variable.entry}{list; entry in dataset yaml for this variable}

\item{name}{character; harmonized name of variable in yaml}

\item{my.theme}{ggplot2 accumulated theme settings}

\item{suppress.reporting}{logical; whether variable report data
should be suppressed}
}
\value{
formatted ggplot for rendering, or nothing if reporting suppressed
}
\description{
Helper function to report plot showing relationship between
self-reported systolic and diastolic blood pressure
}
