% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.excel.problems}
\alias{report.excel.problems}
\title{Helper function to report detected excel problems in variable contents}
\usage{
report.excel.problems(variable.entry, suppress.reporting)
}
\arguments{
\item{variable.entry}{list; entry in dataset yaml for this variable}

\item{suppress.reporting}{logical; whether variable report data
should be suppressed}
}
\description{
Helper function to report detected excel problems in variable contents
}
