% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.factor.summary}
\alias{report.factor.summary}
\title{Helper function to report summary information about uncertain
values in expected categorical variables.}
\usage{
report.factor.summary(variable.entry, name, suppress.reporting)
}
\arguments{
\item{variable.entry}{list; entry in dataset yaml for this variable}

\item{name}{character; harmonized name of variable in yaml}

\item{suppress.reporting}{logical; whether variable report data
should be suppressed}
}
\description{
Helper function to report summary information about uncertain
values in expected categorical variables.
}
\details{
this function has hybrid functionality. it was originally
intended to report entries of factor variables that do not match
defined levels/shared model data. later on, functionality was added
to try to ad hoc harmonize self-reported ancestry labels, and at
that point this function was expanded to provide a detailed summary
of resolution status for such ancestry variables as well.
}
