% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.linked.date}
\alias{report.linked.date}
\title{Helper function to report relationship between an age variable
and its corresponding date.}
\arguments{
\item{data.vec}{Numeric vector of column from phenotype data frame
for this age variable.}

\item{phenotype.data}{Data frame of full phenotype data for selection
of linked variable contents.}

\item{variable.entry}{List entry in dataset yaml for this age variable.}

\item{name}{Character vector harmonized name of variable in yaml.}

\item{my.theme}{ggplot2 accumulated theme settings.}

\item{suppress.reporting}{Logical controlling whether variable report data
should be suppressed.}
}
\value{
Formatted ggplot for rendering, or NULL if reporting suppressed.
}
\description{
This function exclusively functions to emit formatted information
in the phenotype Rmarkdown report. It has been
parametrized out in order to:
\itemize{
\item clean up the report proper;
\item sanitize some of the required hybrid syntax from working with
R content in markdown; and
\item expose the results of this function to testthat
}

You should never need to call this function for any reason.
}
