% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.name.and.code}
\alias{report.name.and.code}
\title{Helper function to report tracking information about variable
name and, for derived variables, code creating the variable.}
\arguments{
\item{variable.entry}{List entry in dataset yaml for this variable.}
}
\description{
This function exclusively functions to emit formatted information
in the phenotype Rmarkdown report. It has been
parametrized out in order to:
\itemize{
\item clean up the report proper;
\item sanitize some of the required hybrid syntax from working with
R content in markdown; and
\item expose the results of this function to testthat
}

You should never need to call this function for any reason.
}
