% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.noncompliant.bp}
\alias{report.noncompliant.bp}
\title{Helper function to report information about blood pressure
variable entries that do not match expected blood pressure
systolic/diastolic reporting format.}
\usage{
report.noncompliant.bp(variable.entry, name, suppress.reporting)
}
\arguments{
\item{variable.entry}{list; entry in dataset yaml for this variable}

\item{name}{character; harmonized name of variable in yaml}

\item{suppress.reporting}{logical; whether variable report data
should be suppressed}
}
\value{
formatted kable for rendering, or nothing if reporting suppressed
}
\description{
Helper function to report information about blood pressure
variable entries that do not match expected blood pressure
systolic/diastolic reporting format.
}
\details{
Many instances of non-compliant blood pressure reporting
formats have been observed in various test datasets. This reporting
information is intended both to record non-compliant data and to
suggest new, "creative" formats that might be supported in patches
to this package
}
