% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.noncompliant.dates}
\alias{report.noncompliant.dates}
\title{Helper function to report summary information about values
in date variables that do not match expected date formats}
\arguments{
\item{variable.entry}{List entry in dataset yaml for this variable.}

\item{name}{Character vector of harmonized name of variable in yaml.}

\item{suppress.reporting}{Logical controlling whether variable report data
should be suppressed.}
}
\value{
Formatted kable for rendering, or NULL if reporting suppressed.
}
\description{
This function exclusively functions to emit formatted information
in the phenotype Rmarkdown report. It has been
parametrized out in order to:
\itemize{
\item clean up the report proper;
\item sanitize some of the required hybrid syntax from working with
R content in markdown; and
\item expose the results of this function to testthat
}

You should never need to call this function for any reason.
}
\details{
Free-text date variables typically contain a wide variety
of input formats. The package heuristically supports a variety of
date formats, but still many observed values are ultimately not
reasonably convertable to year specifications. This function reports
such values, both for reporting purposes and to suggest possible
extensions to the supported date format regex (or not).
}
