% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.noncompliant.numerics}
\alias{report.noncompliant.numerics}
\title{Helper function to report information about entries in
expected-numeric variables that do not match supported
numeric format.}
\arguments{
\item{variable.entry}{List entry in dataset yaml for this variable.}

\item{name}{Character vector of harmonized name of variable in yaml.}

\item{suppress.reporting}{Logical controlling whether variable report data
should be suppressed.}
}
\value{
Formatted kable for rendering, or NULL if reporting suppressed.
}
\description{
This function exclusively functions to emit formatted information
in the phenotype Rmarkdown report. It has been
parametrized out in order to:
\itemize{
\item clean up the report proper;
\item sanitize some of the required hybrid syntax from working with
R content in markdown; and
\item expose the results of this function to testthat
}

You should never need to call this function for any reason.
}
\details{
Substantial processing is applied to string representations
of input data before numeric conversion is attempted. This report
covers entries that fail conversion after such string cleaning
is attempted.
}
