% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.unicode.entries}
\alias{report.unicode.entries}
\title{Helper function to report summary information about Unicode
characters that are not removed by upstream cleaning.}
\usage{
report.unicode.entries(variable.entry, suppress.reporting)
}
\arguments{
\item{variable.entry}{list; entry in dataset yaml for this variable}

\item{suppress.reporting}{logical; whether variable report data
should be suppressed}
}
\value{
formatted kable for rendering, or nothing if reporting suppressed
}
\description{
Helper function to report summary information about Unicode
characters that are not removed by upstream cleaning.
}
\details{
Unicode characters often sneak by the ad hoc conversion
logic used in this package. the linking between Unicode character
and desired ASCII representation has been exposed to config space.
this reporting function includes tracking information about the
observed Unicode character, in the hopes that the user can expand
the mapping table to convert such characters into compliant values.
}
