% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_helper_functions.R
\name{response.greater.than}
\alias{response.greater.than}
\title{Evaluate relative magnitude of variables}
\usage{
response.greater.than(larger.var, smaller.var, allow.equal = TRUE)
}
\arguments{
\item{larger.var}{numeric vector, expected to be larger}

\item{smaller.var}{numeric vector, expected to be smaller}

\item{allow.equal}{logical, whether equality is permitted}
}
\value{
logical vector of same length as input vectors, denoting
whether first vector values are larger than second's
}
\description{
Evaluate relative magnitude of variables
}
