% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_helper_functions.R
\name{response.is.duplicate.of}
\alias{response.is.duplicate.of}
\title{Test concordance of duplicate variables}
\usage{
response.is.duplicate.of(dependent.variable, independent.variable)
}
\arguments{
\item{dependent.variable}{string, name of dependent variable to operate on}

\item{independent.variable}{string, name of independent variable}
}
\value{
a vector of length nrow(phenotype.data) representing the results
of the dependency test
}
\description{
Sanity check to ensure that suspected duplicate variables are concordant.
}
\details{
This function tests that two non-NA values are concordant.  Note
that non-NA/NA mismatches are allowed here, to focus on the most
toxic discordant cases.
}
