% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_helper_functions.R
\name{response.is.greater.than}
\alias{response.is.greater.than}
\title{Test that one variable is greater than, or possibly equal to, the other}
\usage{
response.is.greater.than(
  dependent.variable,
  independent.variable,
  allow.equal = TRUE
)
}
\arguments{
\item{dependent.variable}{Vector for dependent/upstream variable
in the dependency. In this context,
this is the variable that is expected to be greater than the other.
This function is intended to be called from the \code{dependencies}
block of a dataset configuration file, in which case the mapping keys
under \code{variables} (e.g. HW00001) can be called directly.}

\item{independent.variable}{Vector for independent/downstream variable
in the dependency. In this context,
this is the variable that is expected to be less than the other.
This function is intended to be called from the \code{dependencies} variable
block of a dataset configuration file, in which case the mapping keys
under \code{variables} (e.g. HW00001) can be called directly.}

\item{allow.equal}{Logical indicating whether equal values should be considered to pass.}
}
\value{
Logical vector representing the results
of the dependency test between the two provided vectors. Test
represents whether the specified dependency is satisfied.
}
\description{
This is a function designed for use within the user configuration
file, as a convenience to handle structured comparison between
a question and its dependent downstream response.
In the formal parameters, \code{dependent} refers to the logically
downstream question, and \code{independent} refers to the upstream
question on which the \code{dependent} question depends.

Note that the input vectors are specified as \emph{names of variables}
within the input phenotype dataset. At the time the dependencies
are evaluated, the phenotype dataset has had the user-configured
variable names applied as its column headers, and as such the
variables should be referred to by those labels directly.
}
\details{
This dependency test assumes the input vectors are numeric or support '>',
or optionally '>='.
For the purposes of this comparison, NA response in either independent
or dependent variable leads to the dependency test returning success;
the goal of the function is to flag explicit situations when the
dependency definitely fails.
}
\examples{
indep.var <- c(1:5, NA, 5)
dep.var <- c(5:1, 1, NA)
phenotype.data <- data.frame(
  HW00001 = indep.var,
  HW00002 = dep.var
)
dep.test <- response.is.greater.than(
  phenotype.data$HW00002,
  phenotype.data$HW00001,
  TRUE
)
## this function is designed as a utility to be
## deployed in the dependency block of a user configuration
## file. it requires the bindings of the phenotype data
## matrix to be available in the current context
\dontrun{
dep.test <- response.is.greater.than(
  HW00002,
  HW00001,
  TRUE
)
}
## expected: TRUE TRUE TRUE FALSE FALSE TRUE TRUE
}
\seealso{
response.depends.on.yes, response.depends.on.not.na,
response.is.less.than,
response.is.duplicate.of, response.is.computed.bmi
year.is.consistent.with.age
}
