% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_helper_functions.R
\name{response.is.less.than}
\alias{response.is.less.than}
\title{Test that one variable is less than the other}
\usage{
response.is.less.than(dependent.variable, independent.variable)
}
\arguments{
\item{dependent.variable}{string, name of dependent variable to operate on}

\item{independent.variable}{string, name of independent variable}
}
\value{
a vector of length nrow(phenotype.data) representing the results
of the dependency test
}
\description{
Sanity check to ensure that related variables are behaving as expected.
}
\details{
This function tests that one variable is less than another.
}
