% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_header.R
\name{sanitize.header}
\alias{sanitize.header}
\title{Apply transformed variable names to data frame}
\arguments{
\item{df}{Data frame containing phenotype data with untransformed headers.}

\item{var.map}{List output of \code{map.header}, containing input dataset configuration
but with user settings per variable in params sublists.}
}
\value{
Data frame containing input phenotype data with variable name transformation applied.
}
\description{
Header remapping is designed to clean up input format discrepancies such as
non-standard characters, repeat values, etc. The worst we have seen is header emoji;
every time you don't have header emoji, take some time to feel grateful for that.

As with \code{map.header}, this functionality is some of the oldest in the package.
It should not be called outside of the \code{create.phenotype.report} processing chain.
}
\details{
This function takes a data frame and the output of \code{map.header}, and applies
the variable name transformation to the data frame. Note that this must be run
after \code{map.header}, and without any column ordering
changes on the input phenotype data. If the order is changed, this sanitization
will fail silently.
}
\seealso{
map.header
}
