% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_helper_functions.R
\name{year.is.consistent.with.age}
\alias{year.is.consistent.with.age}
\title{Test consistency of age and year of birth}
\usage{
year.is.consistent.with.age(
  dependent.variable,
  independent.variable,
  reference.year,
  acceptable.tolerance
)
}
\arguments{
\item{dependent.variable}{string, name of dependent variable to operate on,
should represent DOB}

\item{independent.variable}{string, name of independent variable, should
represent age}

\item{reference.year}{numeric or string, year from which to subtract year
of birth to calculate age (either as a constant, or the name of a variable
containing numeric values')}

\item{acceptable.tolerance}{numeric, range above or below reported age within
which to call reported and calculated age consistent}
}
\value{
a vector of length nrow(phenotype.data) representing the results
of the dependency test
}
\description{
Sanity check to ensure that reported age in years and calculated age from
reported year of birth are roughly equivalent.  Note that this can also be
used for other age comparisons, e.g. age at diagnosis vs. date of diagnosis.
}
\details{
This function tests that reported and calculated age is within a certain
user-defined tolerance.
}
