% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{add.trailing.metadata}
\alias{add.trailing.metadata}
\title{Add fixed trailing annotations to questionnaire configuration}
\usage{
add.trailing.metadata(out.yaml, dataset.tag, responses)
}
\arguments{
\item{out.yaml}{list; partially constructed variable summary
information for questionnaire}

\item{dataset.tag}{character vector; tag for this dataset}

\item{responses}{character vector; column names from actual
completed survey data}
}
\value{
list; input yaml configuration with additional
variable information appended
}
\description{
Add fixed trailing annotations to questionnaire configuration
}
\details{
Survey responses have fixed output metadata columns
appended. The actual columns included seem to be pulled from
a somewhat amorpheous pool of options; more may need to be
added to the pool.
}
