% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_summary_metrics.R
\name{aggregate.subject.dep.failures}
\alias{aggregate.subject.dep.failures}
\title{Compute per-subject dependency failures counts across all variables}
\usage{
\method{aggregate}{subject.dep.failures}(variable.summary)
}
\arguments{
\item{variable.summary}{list, variable configuration
and summary data}
}
\value{
named integer vector; values are counts of
dependency failures for each subject across all variables,
names are subject IDs
}
\description{
Computes a count for each subject of all dependency failures
across all variables, after all filtering and conversion has
been applied.
}
\details{
This obviously has a bunch of caveats associated with it.
The metric only means as much as the dependencies do.
Large numbers of dependency failures can be introduced
with particular discrepancies (not addressing atypical
"no" responses in some dependent questions, approximate
BMI deviations, atypical N/A values in string variables)
that will inflate these counts without useful meaning.
Discrepancies in early links in some dependency chains
can cause the same subject to fail an entire batch
of related variables (e.g. smoking, alcohol), and
while it's probably appropriate to penalize that datapoint
due to this phenomenon, it's not entirely clear
how many times that penalty should be applied.
}
