% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_summary_metrics.R
\name{aggregate.subjects.wrong.type}
\alias{aggregate.subjects.wrong.type}
\title{Find, for each subject, the number of variables for
which that subject has data of the wrong type.}
\usage{
\method{aggregate}{subjects.wrong.type}(variable.summary)
}
\arguments{
\item{variable.summary}{list, variable configuration and
summary data}
}
\value{
named integer vector; values are counts of non-compliant
variables per subject, names are subject IDs
}
\description{
This aggregates non-compliant subject ID information from
all variables and reduces them to counts of instances
per subject, for downstream tabular reporting.
}
\details{
These counts may be inflated per subject if there is
something specifically wrong with a block of related
variables. This count summary is hopefully a proxy
for subject-specific data corruption, so they can
be removed before downstream analysis.
}
