% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_summary_metrics.R
\name{aggregate.variables.wrong.type}
\alias{aggregate.variables.wrong.type}
\title{Find, for each variable, the number of subjects with entries
not matching the specified type.}
\usage{
\method{aggregate}{variables.wrong.type}(variable.summary)
}
\arguments{
\item{variable.summary}{list, variable configuration and
summary data}
}
\value{
named integer vector; values are counts of non-compliant
subjects per variable, names are variable names
}
\description{
This aggregates counts of non-compliant subjects per variable,
for downstream tabular reporting.
}
\details{
This is distinct from other reporting of non-compliant values,
which are used for screening and variable-specific NA exclusion.
In some instances, inflated counts may reflect either
corruption of the variable, due to technical reasons or changes
in questionnaire content midstream, or potentially configuration
failures in this software.
}
