% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{apply.replacements}
\alias{apply.replacements}
\title{Apply standard replacements for certain deprecated terms}
\usage{
apply.replacements(vec)
}
\arguments{
\item{vec}{character vector}
}
\value{
character vector with replacements applied
}
\description{
Apply standard replacements for certain deprecated terms
}
