% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_phenotype_cleaning.R
\name{apply.type.conversions}
\alias{apply.type.conversions}
\title{Apply type constraints to each variable from yaml config}
\arguments{
\item{phenotype.data}{Data frame of loaded phenotype data with
standardized headers. All columns should be character vectors in
input, as this is the method that casts them out of character vectors.}

\item{variable.summary}{List of per-column summary information
and parameters from dataset configuration.}
}
\value{
List, with first entry 'phenotype.data' a data frame of modified
phenotype input data with conversions applied as appropriate;
and second entry 'variable.summary' a list of modified input configuration
with summary information injected from certain handlers.
}
\description{
Each yaml configuration block should have a "type" entry
denoting what data storage type is expected for this variable.
This function dispatches the appropriate handlers for these
methods, specifically triggering type-specific cleaning in some
cases before casting the values.

If for any reason you don't like the type conversion behaviors
in this package, set the variable type to "string" and then
implement the processing you desire as a derived variable.
}
\details{
This function is a pure internal, and should never be
called outside of the package.
}
\seealso{
load.configuration
}
\keyword{phenotypes}
\keyword{yaml}
