% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{build.variable.data}
\alias{build.variable.data}
\title{Construct variable annotation according to type information
from a SurveyCTO configuration row}
\usage{
build.variable.data(type.value, name.value, label.value, choice.list, varname)
}
\arguments{
\item{type.value}{character vector; entry from "type" column}

\item{name.value}{character vector; entry from "name" column}

\item{label.value}{character vector; entry from "label" column}

\item{choice.list}{list; shared model data for questionnaire}

\item{varname}{character vector; constructed name of variable (e.g. DT00001)}
}
\value{
list; contents of variable summary for this variable, under
variable.summary$variables[\link{varname}]
}
\description{
Construct variable annotation according to type information
from a SurveyCTO configuration row
}
