% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_dependencies.R
\name{check.variable.dependencies}
\alias{check.variable.dependencies}
\title{Apply config-specified R commands as variable sanity checks}
\usage{
check.variable.dependencies(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{data.frame, input phenotype data}

\item{variable.summary}{list, configuration data per variable}
}
\value{
list, modified version of input variable.summary argument
with some reporting information about dependency check success/failure
}
\description{
Certain variables in an input dataset should have defined relationships
between one another. This function will evaluate, in a hopefully safe manner,
R syntax expressions from the config yaml files and report their results
to the per-variable summary data, for reporting in the output html.
}
