% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{collapse.repeats}
\alias{collapse.repeats}
\title{Collapse multiple consecutive instances of specified characters
to a single instance}
\usage{
collapse.repeats(
  df,
  targets = c("\\\\\\\\/", "\\\\."),
  replacements = c("/", ".")
)
}
\arguments{
\item{df}{Data frame of input phenotype data as character vectors.}

\item{targets}{Character vector of targets to collapse into single instances.}

\item{replacements}{Character vector of what the repeats should be collapsed to.}
}
\value{
Data frame containing input entries with specified repeat characters
collapsed into single instances.
}
\description{
Repeated instances of certain types of characters generally indicate
manual input errors, and prevent harmonization of repeated entries.
This function collapses specified characters into single instances of
a corresponding replacement character. In the case of the \code{create.phenotype.report}
processing chain, multiple '.' are collapsed into a single '.', and multiple
entries of any combination of '/' and '\\' are collapsed into a single '/'.
}
\examples{
phenotype.data <- data.frame(
  HW00001 = c("A", "B", "C", "D"),
  HW00002 = c("//", "\\\\\\\\//", "..", "1..1")
)
result <- process.phenotypes:::collapse.repeats(phenotype.data)
}
