% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_files.R
\name{compare.files}
\alias{compare.files}
\title{Compare two files for approximate identity of columns}
\usage{
compare.files(
  first.filename,
  first.id.colname,
  second.filename,
  second.id.colname
)
}
\arguments{
\item{first.filename}{character vector, name of
first file to load and compare}

\item{first.id.colname}{character vector, name
of first file column to treat as subject ID}

\item{second.filename}{character vector, name of
second file to load and compare}

\item{second.id.colnmae}{character vector, name
of second file column to treat as subject ID

note that this function is primary used for its
side effect, emitting content to screen.}
}
\description{
TBD
}
\details{
as a utility function, this compares two datasets
for approximately identical columns, based not on
the header labels but on column values. this is
likely to completely fail until at least minimal
initial processing with the main methods of this
package.
}
