% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_summary_metrics.R
\name{compute.subject.na.count}
\alias{compute.subject.na.count}
\title{Compute per-subject NA count across all variables}
\usage{
compute.subject.na.count(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{data frame, subject data, subjects
as rows, phenotypes as columns}
}
\value{
named integer vector; values are counts of NAs
for each subject across all variables, names are subject IDs
}
\description{
Computes a count for each subject of all NAs across
all variables, after all filtering and conversion has
been applied.
}
\details{
NAs can be present in input data or introduced at various
steps, and the overall count can be indicative of various
phenomena, including expected blockwise NA due to individual
subjects not completing all questionnaire sections.
}
