% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_phenotype_cleaning.R
\name{convert.type}
\alias{convert.type}
\title{Coordinate different function calls for type conversion}
\usage{
convert.type(vec, var.summary, target.type)
}
\arguments{
\item{vec}{character vector, input phenotype content}

\item{var.summary}{list, variable summary entry for this particular variable}

\item{target.type}{character vector, denotes the expected type for each variable;
read from the yaml config}
}
\value{
list, 'phenotype.data' contains converted phenotype information,
'variable.summary' contains input variable summary.
}
\description{
A short list of special handlers are implemented for particular
data conversion problems (e.g. numerics with trailing unit suffixes,
or blood pressure data).  This function dispatches the appropriate
subroutine based on the detected data type.
}
\details{
Each yaml configuration block should have a "type" entry
(if not, will default to unmodified string) denoting what
data storage type is expected for this variable.
}
\seealso{
apply.type.conversion
}
\keyword{phenotypes}
\keyword{yaml}
