% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_phenotype_cleaning.R
\name{convert.variable.specific.na}
\alias{convert.variable.specific.na}
\title{Find additional NA aliases in user config and apply them to
particular variables}
\usage{
convert.variable.specific.na(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{data frame, loaded phenotype data with
standardized headers}

\item{variable.summary}{list, per-column summary information
and parameters from yaml input}
}
\value{
data frame containing phenotype information with
subjects containing specified values set to NA
}
\description{
TBD
}
\details{
User config can optionally contain an 'na-values' sequence
containing entries that should be mapped to NA. This functionality
is only really reliable for factors and strings; numerics may
not match exactly as desired in some cases.
}
