% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{create.config}
\alias{create.config}
\title{Initialize dataset yaml configuration
with basic default entries}
\usage{
create.config(dataset.tag)
}
\arguments{
\item{dataset.tag}{character vector; tag
for dataset}
}
\value{
list representing dataset yaml
with a single placeholder variable
}
\description{
Initialize dataset yaml configuration
with basic default entries
}
