% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{create.derived.variables}
\alias{create.derived.variables}
\title{Apply config-specified R commands to generate derived variables}
\usage{
create.derived.variables(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{data.frame, input phenotype data}

\item{variable.summary}{list, configuration data per variable}
}
\value{
list, modified version of input variable.summary argument
with augmented phenotype dataset, and variable summary supplemented
with copies of the input configuration for the derived variables.
}
\description{
After the first run with a derived variable, that variable will
be present in the output dataset, and in its output data
dictionary. Subsequent loads of the data into this software
will treat the derived variable as a primary variable, and
it will not be recomputed.
}
\details{
Userspace configuration of a dataset can contain a block of
derived variables, which are computed dynamically (after
primary dataset cleaning) from existing variables. This uses
functionality similar to that of dependency checking to allow
arbitrary evaluation of code from the config.

Evaluation of derived variables may use any variable, either
raw or derived, from the user config. Evaluation is conducted
in a restricted environment that prevents modification to the
primary phenotype dataset.
}
