% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_dependencies.R
\name{dependency.failure.handling}
\alias{dependency.failure.handling}
\title{Apply config-specified R commands as variable sanity checks}
\usage{
dependency.failure.handling(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{data.frame, input phenotype data}

\item{variable.summary}{list, configuration data per variable}
}
\value{
data frame, modified version of input phenotype.data argument
with requested subject/variable entries replaced with NA on
dependency failure
}
\description{
TBD
}
\details{
Certain variables in an input dataset should have defined relationships
between one another. This function will evaluate, in a hopefully safe manner,
R syntax expressions from the config yaml files and report their results
to the per-variable summary data, for reporting in the output html.
}
