% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variable_helpers.R
\name{derive.first.degree}
\alias{derive.first.degree}
\title{Derive binary yes/no first-degree relatives from family member data}
\usage{
derive.first.degree(variable.name)
}
\arguments{
\item{variable.name}{string, name of existing variable to operate on}
}
\value{
a vector of length nrow(phenotype.data) representing the
derived variable
}
\description{
This helper function allows one to convert variables that are
lists of family members (e.g. "father,mother,grandmother") to
a yes/no variable that simply reports whether first degree
relatives are included in the list.  We are assuming that the
level of granularity required for downstream analysis is more
along the lines of whether or not first-degree relations are
affected.
}
\details{
Logic to convert family member lists to binary variable indicating
whether or not first degree relatives are affected.  Handled through
rlang::eval_tidy, which provides an environment and data mask to
prevent overwriting of data in memory (see functions in
derived_variables.R).
}
