% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{emit.markdown.header}
\alias{emit.markdown.header}
\title{Helper function to emit formatted markdown header summarizing
data cleaning process}
\arguments{
\item{dataset.yaml}{Character vector name of input dataset yaml file.}

\item{total.initial.sample.size}{Numeric number of lines in initial input data file.}

\item{consent.inclusion.file}{Character or NULL. Name of input file containing
consent approved subject list, or NULL if none provided.}

\item{subjects.consent.yes}{Numeric; if consent.inclusion.file is not
NULL, the number of subjects in the input dataset that were present in the
input consent approved subject list.}

\item{consent.exclusion.file}{Character or NULL. Name of input file containing
consent rejected subject list, or NULL if none provided.}

\item{subjects.consent.no}{Numeric; if consent.exclusion.file is not
NULL, the number of subjects in the input dataset that were present in the
input consent rejected subject list.}

\item{subjects.ambiguous.consent}{Numeric; if consent.inclusion.file and
consent.exclusion.file are both not NULL, the number of subjects in the
input dataset that were absent from both consent lists.}

\item{subjects.excluded.for.age}{Numeric number of subjects with reported
age less than the minimum value specified in dataset yaml/globals/min_age_for_inclusion.}

\item{min.age.for.inclusion}{Numeric entry from dataset yaml/globals/min_age_for_inclusion.}

\item{na.subject.id.count}{Numeric number of subjects with invalid subject ID entries.}
}
\value{
Character formatted header content.
}
\description{
This function exclusively functions to emit formatted information
for the top of the phenotype Rmarkdown report. It has been
parametrized out in order to:
\itemize{
\item clean up the report proper;
\item sanitize some of the required hybrid syntax from working with
R content in markdown; and
\item expose the results of this function to testthat
}

You should never need to call this function for any reason.
}
