% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{evaluate.derived.expressions}
\alias{evaluate.derived.expressions}
\title{TODO}
\usage{
evaluate.derived.expressions(phenotype.data, derived.exprs)
}
\arguments{
\item{phenotype.data}{data.frame, input phenotype data}

\item{derived.exprs}{list, parsed expressions}
}
\value{
TBD
}
\description{
TBD
}
\details{

}
