% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_phenotype_cleaning.R
\name{exclude.by.age}
\alias{exclude.by.age}
\title{Exclude subjects below a certain user-defined age threshold}
\usage{
exclude.by.age(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{data frame, loaded phenotype data with
standardized headers; all columns should be character vectors}

\item{variable.summary}{list, per-column summary information
and parameters from yaml input}
}
\value{
list, 'phenotype.data' contains phenotype information with
subjects below the age threshold excluded, 'variable.summary' contains
input variable summary with information about excluded subjects.
}
\description{
This function will exclude subjects whose age, as defined with the
\code{subject_age: true} flag in the config, falls below the threshold.
This is useful for excluding minors, for example.
}
