% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_phenotype_cleaning.R
\name{exclude.by.missing.subject.id}
\alias{exclude.by.missing.subject.id}
\title{Exclude subjects for whom no ID has been recorded}
\usage{
exclude.by.missing.subject.id(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{data frame, loaded phenotype data with
standardized headers; all columns should be character vectors}

\item{variable.summary}{list, per-column summary information
and parameters from yaml input}
}
\value{
list, 'phenotype.data' contains phenotype information with
subjects without a subject ID excluded, 'variable.summary' contains
input variable summary with information about excluded subjects.
}
\description{
This function excludes subjects where their ID is NA, as
we can't really do anything with the phenotype data without
being able to link it to genotypes via IDs.
}
