% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{exclude.excel.failures}
\alias{exclude.excel.failures}
\title{Set Excel error messages to NA and flag them as special problems}
\usage{
exclude.excel.failures(phenotype.data, variable.summary)
}
\arguments{
\item{phenotype.data}{data.frame, input phenotype data}

\item{variable.summary}{list, per-variable summary information and configuration data}
}
\value{
list, first entry the input phenotype data with Excel error codes set to NA;
second entry the variable summary list with injected reporting information about
any Excel errors found for each variable
}
\description{
The function \code{remove.nonword.chars} will remove the leading and trailing characters
from default Excel error codes, so for the least ambiguity, it should be called
after Excel errors are already removed.
}
\details{
This function removes "#error!" and "#value!" Excel error codes and reports
the count of any such removals to the per-variable summary list, for inclusion
in the downstream report.
}
