% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_dependencies.R
\name{find.subject.id.index}
\alias{find.subject.id.index}
\title{Locate a variable in the dataset flagged as subject ID}
\usage{
find.subject.id.index(variable.summary)
}
\arguments{
\item{variable.summary}{list, input configuration data}
}
\value{
integer, index of subject ID column in input data
}
\description{
the input yaml specification allows a single variable
to be flagged as "subject_id: true" for context-specific
behaviors.
}
\details{
if this column is not located, stop() is called. in the
future, this column's presence and uniqueness will be
enforced upstream with yaml consistency checking.
}
