% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{handle.multiple.levels}
\alias{handle.multiple.levels}
\title{Populate variables list with multiple response onehot variables}
\usage{
handle.multiple.levels(
  choice.list,
  shared.model,
  varname,
  name.value,
  label.value,
  res
)
}
\arguments{
\item{choice.list}{data.frame; xlsx choices tab}

\item{shared.model}{character; name of categorical model from xlsx choices tab}

\item{varname}{character; harmonized variable name for multiple response group}

\item{name.value}{character; name entry for variable from xlsx survey tab}

\item{label.value}{character; label entry for variable from xlsx survey tab}

\item{res}{list; output yaml list in progress of being built}
}
\value{
list; input res variable with variables extended with new entries
}
\description{
Populate variables list with multiple response onehot variables
}
