% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{handle.repeat.variables}
\alias{handle.repeat.variables}
\title{Create variable annotations for repeat variable blocks}
\usage{
handle.repeat.variables(
  out.yaml,
  cur.varname,
  name.value,
  label.value,
  survey,
  dataset.tag,
  responses,
  i
)
}
\arguments{
\item{out.yaml}{list; partially constructed variable summary}

\item{cur.varname}{character vector; constructed name of repeat start variable}

\item{name.value}{character vector; name column entry for repeat start variable}

\item{label.value}{character vector; label column entry for repeat start variable}

\item{survey}{data.frame; survey tab from SurveyCTO configuration xlsx file}

\item{dataset.tag}{character vector; tag for current dataset}

\item{responses}{character vector; column names of completed questionnaire csv}

\item{i}{index of repeat start variable in survey configuration table}
}
\value{
list; input variable yaml configuration with repeat variable block
data added as entry "out.yaml", and incremented global counter as entry "i"
}
\description{
Create variable annotations for repeat variable blocks
}
