% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ancestry_labels.R
\name{harmonize.ancestry}
\alias{harmonize.ancestry}
\title{Use external label dictionary to harmonize free-text
ancestry labels}
\usage{
harmonize.ancestry(
  phenotype.data,
  variable.summary,
  ancestry.source = "nigeria",
  best.match.threshold = 0.9,
  best.match.discernment = 0.75
)
}
\arguments{
\item{phenotype.data}{data frame, input phenotype data}

\item{variable.summary}{list, variable configuration data}

\item{ancestry.source}{character vector, name of data
source to pull known labels from (e.g. "nigeria"); or
a file in the format expected by load.ancestry.linker}

\item{best.match.threshold}{numeric, on [0,1], denoting how high
the best match similarity must be to permit a call}

\item{best.match.discernment}{numeric, on [0,1] denoting the proportion
of the top match the next best match must be less than (assuming the
next best match does not map to the same harmonized label) in order
to accept the top match}
}
\value{
list; first entry is modified phenotype dataset
with ancestry labels updated, second entry is modified
variable summary with ancestry label update metrics added
}
\description{
Ancestry information has been collected as free-text entries
with little effort at formal harmonization. Though it may be
impossible to perfectly recapture the underlying data due
to high string similarity and short length of distinct labels,
it's worth a shot.
}
\details{
TBD
}
\seealso{
load.ancestry.linker
}
