% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ancestry_labels.R
\name{load.ancestry.linker}
\alias{load.ancestry.linker}
\title{Read ancestry linker data and format it for use in heuristic
ancestry mapping.}
\usage{
load.ancestry.linker(filename)
}
\arguments{
\item{filename}{Character vector name of file containing
structured linker data.}
}
\value{
Named character vector of ancestries;
names are unique expected inputs in phenotype dataset,
values are harmonized labels to be applied in the output.
}
\description{
Free-text self-reported ancestry is expected to be heterogeneous
and complicated. For use in genomics contexts, the labels should
be harmonized, at the very least applying single standardized
group labels to groups with multiple possible (typically quite
similar) names.
}
\details{
See the help documentation of harmonize.ancestry for a discussion
of the motivation behind this and related functions.

Note that, because this function is goofy nonsense,
triple colons are not allowed in any ancestry group name
for the time being.
}
\examples{
test.file <- system.file("external", "nigeria.ancestry.tsv", package = "process.phenotypes")
anc.data <- process.phenotypes:::load.ancestry.linker(test.file)
}
\seealso{
harmonize.ancestry, harmonize.ancestry.from.linker,
weak.ancestry.match
}
