% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_configuration.R
\name{load.configuration}
\alias{load.configuration}
\title{Load yaml configuration from file, integrating global models}
\usage{
load.configuration(dataset.parameter.filename, global.parameter.filename)
}
\arguments{
\item{dataset.parameter.filename}{character vector, filename
of dataset-specific variable parameters}

\item{global.parameter.filename}{character vector, filename
of global shared model specifications}
}
\value{
list of lists, containing parsed yaml configuration
information
}
\description{
Load cleaning configurations from yaml file and apply
shared models.
}
\keyword{yaml}
