% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{make.lowercase}
\alias{make.lowercase}
\title{Basic global cleanup of entries in a phenotype data frame}
\usage{
make.lowercase(df)
}
\arguments{
\item{df}{data frame, input phenotype content}
}
\value{
modified version of input with values cleaned as described
above
}
\description{
Converts all characters to lower case
}
\details{
Performs the following data modifications to non-numeric values:
\itemize{
\item convert all characters to lower case
}
}
