% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{make.lowercase}
\alias{make.lowercase}
\title{Basic global cleanup of entries in a phenotype data frame}
\usage{
make.lowercase(df)
}
\arguments{
\item{df}{data frame, input phenotype content}
}
\value{
modified version of input with values cleaned as described
above
}
\description{

}
\details{
Performs the following data modifications to non-numeric values:
\itemize{
\item convert all characters to lower case
}
}
\examples{
phenotype.data <- data.frame(
  c("WeiRd CaPs", "Trailing ", "NA"),
  c(";something", "Not Applicable", "Too   muchspace")
)
colnames(phenotype.data) <- c("col1", "col2")
phenotype.data <- make.lowercase(phenotype.data)
}
