% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_header.R
\name{map.header}
\alias{map.header}
\title{Create and map structured variable names to replace arbitrary input variable names}
\usage{
map.header(df, dataset.tag)
}
\arguments{
\item{df}{data.frame, phenotype dataframe with untransformed headers}

\item{dataset.tag}{character vector, unique string tag for this dataset}
}
\value{
list, mapping new variable names to lists; eventually meant to have
more entries but will for now just have a original.name entry for mapping back
to the raw variable names
}
\description{

}
\details{
Input data frame may have any format variable names stored as column
headers; this function maps those to standardized format "TAG#####" names.
The new names are reported as names(result), and the old variable names
corresponding to those names are reported as result${newname}$original.name.
}
\examples{
df <- data.frame(rnorm(100), runif(100))
colnames(df) <- c("Human height lol", "y")
map.header(df, "mytag")
}
\seealso{
sanitize.header
}
\keyword{phenotypes}
