% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_header.R
\name{map.header}
\alias{map.header}
\title{Create and map structured variable names to replace arbitrary input variable names}
\usage{
map.header(df, dataset.tag, config.data, force.header.mapping = FALSE)
}
\arguments{
\item{df}{data.frame, phenotype dataframe with untransformed headers}

\item{dataset.tag}{character vector, unique string tag for this dataset}

\item{config.data}{list, yaml configuration data per variable}

\item{force.header.mapping}{logical, whether you want to be kinda foolish
and allow desync between config variable names and dataset header data.
this safety check was implemented in part due to the possible presence of duplicate
column header names, and turning it off can potentially have catastrophically
bad effects if your configuration file is malformed.}
}
\value{
list, mapping new variable names to lists; eventually meant to have
more entries but will for now just have a original.name entry for mapping back
to the raw variable names
}
\description{
Since the addition of yaml configuration, this function additionally grabs
any available configuration information for each variable and stores it as
a "params" list entry per variable.
}
\details{
Input data frame may have any format variable names stored as column
headers; this function maps those to standardized format "TAG#####" names.
The new names are reported as names(result), and the old variable names
corresponding to those names are reported as result${newname}$original.name.
}
\seealso{
sanitize.header
}
\keyword{phenotypes}
