% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{normalize.missing.values}
\alias{normalize.missing.values}
\title{Basic global cleanup of entries in a phenotype data frame}
\usage{
normalize.missing.values(df)
}
\arguments{
\item{df}{data frame, input phenotype content}
}
\value{
modified version of input with values cleaned as described
above
}
\description{

}
\details{
Performs the following data modifications to non-numeric values:
\itemize{
\item harmonize na, nan, not applicable values (unknowns?  blanks?)
}
}
\examples{
phenotype.data <- data.frame(
  c("WeiRd CaPs", "Trailing ", "NA"),
  c(";something", "Not Applicable", "Too   muchspace")
)
colnames(phenotype.data) <- c("col1", "col2")
phenotype.data <- normalize.missing.values(phenotype.data)
}
