% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_phenotype_cleaning.R
\name{parse.date}
\alias{parse.date}
\title{Convert values that should be dates to year}
\usage{
parse.date(vec, var.summary)
}
\arguments{
\item{vec}{Character vector of input phenotype content.}

\item{var.summary}{List of variable summary entry for this particular variable.}
}
\value{
Modified version of input vector with values cleaned as described above.
}
\description{
Given string vector with malformed date entries, attempt
to force conversion to YYYY.
}
\details{
Extracts two or four digit year values from dates of several formats,
assigns likely century (20 or 19), and replaces the original date
with just the four-digit year value.  Also removes extremely low
likely erroneous year values.

With new data, new support for the following data formats:
\itemize{
\item YYYY
\item YYYY-##-##
\item any amount of text; a space, -, or /; and year as YY or YYYY
\item any of the 12 Gregorian calendar months without typos, an optional comma, and then YYYY
}
}
\examples{
data <- c("1987", "1999-12-11")
var.summary <- list()
result <- process.phenotypes:::parse.date(data, var.summary)
}
