% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{parse.surveycto}
\alias{parse.surveycto}
\title{Convert SurveyCTO configuration data into yaml configuration for this package}
\usage{
parse.surveycto(
  in.form.filename,
  in.response.filename,
  dataset.tag,
  out.yaml.filename,
  out.shared.models,
  subject.id.name = "pid",
  age.name = "age"
)
}
\arguments{
\item{in.form.filename}{character vector; name of xlsx survey
configuration file from SurveyCTO}

\item{in.response.filename}{character vector; name of csv
response data from SurveyCTO in wide format}

\item{dataset.tag}{character vector; tag for this dataset,
which is used as prefix for synthetic variable names}

\item{out.yaml.filename}{character vector; destination filename
for variable configuration data}

\item{out.shared.models}{character vector; destination filename
for shared categorical model data for this questionnaire}

\item{subject.id.name}{character vector; name of variable containing subject ID}

\item{age.name}{character vector; name of variable containing subject age}
}
\description{
Convert SurveyCTO configuration data into yaml configuration for this package
}
