% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveycto.R
\name{populate.choices}
\alias{populate.choices}
\title{Render SurveyCTO configuration "choices" tab
as a shared_models style yaml list}
\usage{
populate.choices(df, survey.type)
}
\arguments{
\item{df}{data.frame, input choices tab from SurveyCTO spreadsheet
with column headers "list_name", "value", and "label"
at least present}

\item{survey.type}{character vector; type column entries
from "survey" tab of CTO form configuration}
}
\value{
list of choice information as shared_models
yaml list, with all variables configured as categoricals
}
\description{
Render SurveyCTO configuration "choices" tab
as a shared_models style yaml list
}
