% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phenotypes.R
\docType{package}
\name{process.phenotypes}
\alias{process.phenotypes}
\title{process.phenotypes: automated phenotype standardization and reporting}
\description{
A package for reproducible phenotype data cleaning and quality control.
}
\seealso{
\itemize{
\item \url{https://54geneprocessphenotypes.readthedocs.io/en/latest/}
\item \url{https://gitlab.com/data-analysis5/phenotypes/process.phenotypes}
}
}
