% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{process.unicode.characters}
\alias{process.unicode.characters}
\title{Convert sporadic Unicode-only characters into lesser equivalents}
\usage{
process.unicode.characters(phenotype.data)
}
\arguments{
\item{phenotype.data}{data.frame, input phenotype data}
}
\value{
data.frame input data with Unicode characters converted
into more manageable equivalents.
}
\description{
Note that there's something particularly strange going on with
Excel's "=#ERROR!" code, as in one instance it's getting converted
into, of all things, an emoji.

The functionality of this method is now controlled by a plaintext
configuration file located in inst/unicode_pattern_replacements.tsv.
}
\details{
Sporadic Unicode characters are being inserted into the dataset by
upstream text editors. This function attempts to take them right
back out again. Note that the replacements should be lowercase
to be consistent with other string processing functions.
}
