% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{reformat.factor}
\alias{reformat.factor}
\title{Convert character vector to factor, collapsing certain levels}
\usage{
reformat.factor(vec, variable.summary)
}
\arguments{
\item{vec}{character vector, input string phenotype data}

\item{variable.summary}{list, summary information for a given variable}
}
\value{
list, 'phenotype.data' contains converted phenotype information,
'variable.summary' contains input variable summary and possibly information
about invalid factor levels converted to NA.
}
\description{
Reformats character vector to factor
}
\details{
Converts character vector to factor based on level specification in yaml config.
This will report entries that are converted to NA without first being set to NA intentionally.
}
