% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{reformat.numerics}
\alias{reformat.numerics}
\title{Aggressively convert malformed numeric vector}
\usage{
reformat.numerics(vec, var.summary)
}
\arguments{
\item{vec}{character vector, input phenotype content}

\item{var.summary}{list, variable summary entry for this particular variable}
}
\value{
modified version of input with values cleaned as described above
}
\description{
Remove trailing non-numerics to allow conversion to numeric
}
\details{
Given string vector with malformed numeric entries, attempt
to force conversion to numerics by stripping common problems.
}
