% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{reformat.numerics}
\alias{reformat.numerics}
\title{Attempt to convert somewhat malformed numeric vectors to actual numerics}
\usage{
reformat.numerics(vec, var.summary)
}
\arguments{
\item{vec}{Character vector of input candidate numeric data.}

\item{var.summary}{List of input variable configuration data for this data vector.}
}
\value{
List, with first entry \code{phenotype.data} modified input data vector with
numeric conversion applied, and second entry \code{variable.summary} the input
configuration data with invalid numeric entries annotated for later reporting.
}
\description{
Given a string vector with malformed numeric entries, this function attempt
to force conversion to numerics by stripping common problems.
}
\details{
This conversion is potentially \emph{very} aggressive for certain types of inputs.
As such, anything that fails conversion is recorded and reported out as
a table in the \code{create.phenotype.report} cleaning report. Each time there's
anything in such a table for any variable, the values should be evaluated
and inspected for possible issues, or potentially any requirements for
updating the accepted numeric format definition.
}
\examples{
data <- c("100", "120.1", "1.5 meters")
var.summary <- list()
results <- process.phenotypes:::reformat.numerics(data, var.summary)
}
