% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_invalid_columns.R
\name{remove.invalid.columns}
\alias{remove.invalid.columns}
\title{Remove bad columns from a phenotype data frame}
\usage{
remove.invalid.columns(df)
}
\arguments{
\item{df}{data frame, input phenotype content}
}
\value{
modified version of input with certain bad columns
removed
}
\description{
Removes columns under the following criteria:
\itemize{
\item column name is NA
\item column name is empty string
}
}
\examples{
\dontrun{
phenotype.data <- data.frame(1:3, 4:6, 7:9)
colnames(phenotype.data) <- c("", NA, "ok name")
phenotype.data <- remove.invalid.columns(phenotype.data)
}
}
