% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{remove.nonword.chars}
\alias{remove.nonword.chars}
\title{Basic global cleanup of entries in a phenotype data frame}
\usage{
remove.nonword.chars(df, variable.summary)
}
\arguments{
\item{df}{data frame, input phenotype content}

\item{variable.summary}{list, variable-specific configuration data}
}
\value{
modified version of input with values cleaned as described
above
}
\description{
Removes preceding and trailing non-alphanumeric characters
}
\details{
Performs the following data modifications to non-numeric values:
\itemize{
\item remove trailing and preceding non-alphanumerics
}
}
