% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_vector_cleanup.R
\name{remove.whitespace}
\alias{remove.whitespace}
\title{Basic global cleanup of entries in a phenotype data frame}
\usage{
remove.whitespace(df)
}
\arguments{
\item{df}{data frame, input phenotype content}
}
\value{
modified version of input with values cleaned as described
above
}
\description{
Removes all trailing and preceding whitespaces; collapses
multiple consecutive whitespaces
}
\details{
Performs the following data modifications to non-numeric values:
\itemize{
\item remove all trailing and preceding whitespaces and collapse
consecutive whitespaces
}
}
