% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.content.summary}
\alias{report.content.summary}
\title{Helper function to report information about observed entries
in text-like variables}
\usage{
report.content.summary(
  phenotype.data,
  var.summary,
  unique.var.value.inc.prop,
  name,
  variable.pretty.name,
  suppress.reporting
)
}
\arguments{
\item{phenotype.data}{data.frame; full phenotype data for variable
selection}

\item{var.summary}{table; contingency data about observations
of values in variable}

\item{unique.var.value.inc.prop}{numeric; the proportion threshold
above which variables with too many unique values will not have
their value contingency data reported}

\item{name}{character; harmonized name of variable in yaml}

\item{variable.pretty.name}{character; human-legible name of
variable with more helpful description}

\item{suppress.reporting}{logical; whether variable report data
should be suppressed}
}
\value{
formatted kable for rendering, or nothing if reporting suppressed
}
\description{
Helper function to report information about observed entries
in text-like variables
}
