% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.dependencies}
\alias{report.dependencies}
\title{Helper function to report summary information about encoded
dependency relationships between variables.}
\usage{
report.dependencies(phenotype.data, variable.summary, name, suppress.reporting)
}
\arguments{
\item{phenotype.data}{data.frame; full phenotype data for selection
of linked variable contents}

\item{variable.summary}{list; input dataset config yaml}

\item{name}{character; harmonized name of variable in yaml}

\item{suppress.reporting}{logical; whether variable report data
should be suppressed}
}
\value{
list; entries 'contingency' and 'cross', for kables describing
contingency table of results and cross-variable comparisons; or null,
if reporting suppressed or the relevant features not configured
}
\description{
Helper function to report summary information about encoded
dependency relationships between variables.
}
\details{
this function handles all reporting of variable dependencies
from the input config yaml dependency blocks. currently, the report
does not contain direct information about the downstream effects of
exclude_on_error or exclude_all_on_error directives. this may
be added in future iterations of the report.
}
