% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.linked.date}
\alias{report.linked.date}
\title{Helper function to report relationship between an age variable
and its corresponding date}
\usage{
report.linked.date(
  data.vec,
  phenotype.data,
  variable.entry,
  name,
  my.theme,
  suppress.reporting
)
}
\arguments{
\item{data.vec}{numeric vector; column from phenotype dataframe
for this age variable}

\item{phenotype.data}{data.frame; full phenotype data for selection
of linked variable contents}

\item{variable.entry}{list; entry in dataset yaml for this age variable}

\item{name}{character; harmonized name of variable in yaml}

\item{my.theme}{ggplot2 accumulated theme settings}

\item{suppress.reporting}{logical; whether variable report data
should be suppressed}
}
\description{
Helper function to report relationship between an age variable
and its corresponding date
}
