% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.name.and.code}
\alias{report.name.and.code}
\title{Helper function to report tracking information about variable
name and, for derived variables, code creating the variable}
\usage{
report.name.and.code(variable.entry)
}
\arguments{
\item{variable.entry}{list; entry in dataset yaml for this variable}
}
\description{
Helper function to report tracking information about variable
name and, for derived variables, code creating the variable
}
