% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.noncompliant.bp}
\alias{report.noncompliant.bp}
\title{Helper function to report information about blood pressure
variable entries that do not match expected blood pressure
systolic/diastolic reporting format}
\arguments{
\item{variable.entry}{List entry in dataset yaml for this variable.}

\item{name}{Character vector of harmonized name of variable in yaml.}

\item{suppress.reporting}{Logical controlling whether variable report data
should be suppressed.}
}
\value{
Formatted kable for rendering, or NULL if reporting suppressed.
}
\description{
This function exclusively functions to emit formatted information
in the phenotype Rmarkdown report. It has been
parametrized out in order to:
\itemize{
\item clean up the report proper;
\item sanitize some of the required hybrid syntax from working with
R content in markdown; and
\item expose the results of this function to testthat
}

You should never need to call this function for any reason.
}
\details{
Many instances of non-compliant blood pressure reporting
formats have been observed in various test datasets. This reporting
information is intended both to record non-compliant data and to
suggest new, "creative" formats that might be supported in patches
to this package.
}
