% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.noncompliant.dates}
\alias{report.noncompliant.dates}
\title{Helper function to report summary information about values
in date variables that do not match expected date formats.}
\usage{
report.noncompliant.dates(variable.entry, name, suppress.reporting)
}
\arguments{
\item{variable.entry}{list; entry in dataset yaml for this variable}

\item{name}{character; harmonized name of variable in yaml}

\item{suppress.reporting}{logical; whether variable report data
should be suppressed}
}
\description{
Helper function to report summary information about values
in date variables that do not match expected date formats.
}
\details{
free-text date variables typically contain a wide variety
of input formats. the package heuristically supports a variety of
date formats, but still many observed values are ultimately not
reasonably convertable to year specifications. this function reports
such values, both for reporting purposes and to suggest possible
extensions to the supported date format regex (or not).
}
