% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper_functions.R
\name{report.noncompliant.numerics}
\alias{report.noncompliant.numerics}
\title{Helper function to report information about entries in
expected-numeric variables that do not match supported
numeric format.}
\usage{
report.noncompliant.numerics(variable.entry, name, suppress.reporting)
}
\arguments{
\item{variable.entry}{list; entry in dataset yaml for this variable}

\item{name}{character; harmonized name of variable in yaml}

\item{suppress.reporting}{logical; whether variable report data
should be suppressed}
}
\value{
formatted kable for rendering, or nothing if reporting suppressed
}
\description{
Helper function to report information about entries in
expected-numeric variables that do not match supported
numeric format.
}
\details{
substantial processing is applied to string representations
of input data before numeric conversion is attempted. this report
covers entries that fail conversion after such string cleaning
is attempted.
}
