% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_helper_functions.R
\name{response.depends.on.not.na}
\alias{response.depends.on.not.na}
\title{Test dependency of one variable on non-NA response of another}
\usage{
response.depends.on.not.na(
  dependent.variable,
  independent.variable,
  allow.no = FALSE,
  no.aliases = c("no"),
  additional.na.levels = c()
)
}
\arguments{
\item{dependent.variable}{string, name of dependent variable to operate on}

\item{independent.variable}{string, name of independent variable}

\item{allow.no}{logical, whether to allow "no" answers as equivalent to NA
in the dependent variable}

\item{no.aliases}{character vector, values to be interpreted as
"no" responses in the dependent variable}

\item{additional.na.levels}{vector, define alternative values to be treated
as NAs in the independent variable, e.g. "0 times"}
}
\value{
a vector of length nrow(phenotype.data) representing the results
of the dependency test
}
\description{
Sanity check to ensure that related variables are behaving as expected.
}
\details{
This function tests two variables where one result depends
on any non-NA result in the other.
}
