% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_helper_functions.R
\name{response.depends.on.yes}
\alias{response.depends.on.yes}
\title{Test dependency of one variable on "yes" response of another}
\usage{
response.depends.on.yes(
  dependent.variable,
  independent.variable,
  allow.no = FALSE,
  additional.na.levels = c()
)
}
\arguments{
\item{dependent.variable}{string, name of dependent variable to operate on}

\item{independent.variable}{string, name of independent variable}

\item{allow.no}{logical, whether to allow "no" answers as equivalent to NA}

\item{additional.na.levels}{vector, define alternative values to be treated
as NAs, e.g. "0 times"}
}
\value{
a vector of length nrow(phenotype.data) representing the results
of the dependency test
}
\description{
Sanity check to ensure that related variables are behaving as expected.
}
\details{
This function tests two variables where one result depends
on a "yes" result in the other.
}
