% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_helper_functions.R
\name{response.depends.on.yes}
\alias{response.depends.on.yes}
\title{Test dependency of one variable on "yes" response of another}
\usage{
response.depends.on.yes(
  dependent.variable,
  independent.variable,
  yes.aliases = c("yes"),
  independent.na.aliases = character(),
  dependent.na.aliases = character()
)
}
\arguments{
\item{dependent.variable}{string, name of dependent variable to operate on}

\item{independent.variable}{string, name of independent variable}

\item{yes.aliases}{character vector, values to be interpreted as "yes"
in the independent variable}

\item{independent.na.aliases}{character vector, define alternative values to be treated
as NAs in the independent variable, e.g. "0 times"}

\item{dependent.na.aliases}{character vector, define alternative values to be treated
as NAs in the dependent variable, e.g. "0 times"}
}
\value{
a vector of length nrow(phenotype.data) representing the results
of the dependency test
}
\description{
Sanity check to ensure that related variables are behaving as expected.
}
\details{
This function tests two variables where one result depends
on a "yes" result in the other.
}
