% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_helper_functions.R
\name{response.is.computed.bmi}
\alias{response.is.computed.bmi}
\title{Test consistency of BMI with height and weight}
\usage{
response.is.computed.bmi(bmi, weight, height, tolerance)
}
\arguments{
\item{bmi}{numeric vector, reported BMI}

\item{weight}{numeric vector, reported weight}

\item{height}{numeric vector, reported height}

\item{tolerance}{numeric, acceptable deviation
between computed and reported BMI}
}
\value{
a vector of length nrow(phenotype.data) representing
the results of the dependency test
}
\description{
This dependency will enforce BMI ~= weight/height^2,
with weight in kilograms and height in meters, within
a certain absolute tolerance.
}
\details{
Large numbers of deviations are expected when the
reported BMI has been arbitrarily entered by respondents
and not computed directly from the source data.
}
