% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_helper_functions.R
\name{response.is.duplicate.of}
\alias{response.is.duplicate.of}
\title{Test concordance of duplicate variables.}
\usage{
response.is.duplicate.of(dependent.variable, independent.variable)
}
\arguments{
\item{dependent.variable}{Vector for dependent/upstream variable
in the dependency.
This function is intended to be called from the \code{dependencies}
block of a dataset configuration file, in which case the mapping keys
under \code{variables} (e.g. HW00001) can be called directly.}

\item{independent.variable}{Vector name of independent/downstream variable
for the dependency.
This function is intended to be called from the \code{dependencies}
block of a dataset configuration file, in which case the mapping keys
under \code{variables} (e.g. HW00001) can be called directly.}
}
\value{
Logical vector representing the results
of the dependency test between the two provided vectors. Test
represents whether the specified dependency is satisfied.
}
\description{
This is a function designed for use within the user configuration
file, as a convenience to handle structured comparison between
a question and its dependent downstream response.
In the formal parameters, \code{dependent} refers to the logically
downstream question, and \code{independent} refers to the upstream
question on which the \code{dependent} question depends.

Note that the input vectors are specified as \emph{names of variables}
within the input phenotype dataset. At the time the dependencies
are evaluated, the phenotype dataset has had the user-configured
variable names applied as its column headers, and as such the
variables should be referred to by those labels directly.
}
\details{
This dependency test only assumes the input vectors can be compared with
the equality operator.
This function tests that two non-NA values are concordant.  Note
that non-NA/NA mismatches are allowed here, to focus on the most
toxic discordant cases.
}
\examples{
indep.var <- c("yes", "sometimes", "no response", NA, "never")
dep.var <- c("yes", "sometimes", NA, "sometimes", "sometimes")
phenotype.data <- data.frame(
  HW00001 = indep.var,
  HW00002 = dep.var
)
dep.test <- response.is.duplicate.of(
  phenotype.data$HW00002,
  phenotype.data$HW00001
)
## this function is designed as a utility to be
## deployed in the dependency block of a user configuration
## file. it requires the bindings of the phenotype data
## matrix to be available in the current context
\dontrun{
dep.test <- response.is.duplicate.of(
  HW00002,
  HW00001
)
}
## expected: TRUE TRUE TRUE TRUE FALSE
}
\seealso{
response.depends.on.yes, response.depends.on.not.na,
response.is.less.than, response.is.greater.than,
response.is.computed.bmi
year.is.consistent.with.age
}
