% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_header.R
\name{sanitize.header}
\alias{sanitize.header}
\title{Apply transformed variable names to data frame}
\usage{
sanitize.header(df, var.map)
}
\arguments{
\item{df}{data frame, phenotype dataframe with untransformed headers}

\item{var.map}{output of \code{map.header} for df}
}
\value{
input data frame with variable name transformation applied
}
\description{

}
\details{
This function takes a data frame and the output of \code{map.header}, and applies
the variable name transformation to the data frame.

Note that this must be run after \code{map.header}, and without any column ordering
changes on the input phenotype data. If the order is changed, this sanitization
will fail silently.
}
\examples{
df <- data.frame(rnorm(100), runif(100))
colnames(df) <- c("Human height lol", "y")
mapped.values <- map.header(df, "mytag")
df.transformed <- sanitize.header(df, mapped.values)
}
\seealso{
map.header
}
\keyword{phenotypes}
