% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_configuration.R
\name{write.configuration}
\alias{write.configuration}
\title{Report final configuration data to file}
\usage{
write.configuration(variable.summary, out.filename)
}
\arguments{
\item{variable.summary}{list, configuration data from
create.phenotype.report run}

\item{out.filename}{character vector, name of file
to which to write yaml-format configuration data}
}
\description{
Configuration data from input config.yaml are modified on load,
and some information is lost. Before reporting, the variable
summary information from the run needs to be somewhat reformatted
into a version that is compatible with the input format.
}
