% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_formats.R
\name{write.output.formats}
\alias{write.output.formats}
\title{Report final phenotype dataset after processing}
\usage{
write.output.formats(
  phenotype.data,
  variable.summary,
  out.prefix,
  write.tsv,
  write.stata,
  write.spss,
  write.sas,
  write.yaml
)
}
\arguments{
\item{phenotype.data}{data.frame, stored phenotype dataset with all cleaning
and filtering applied}

\item{variable.summary}{list, loaded and modified configuration data at
the end of a cleaning run}

\item{out.prefix}{character vector, prefix for all requested output files.
will have the relevant format-specific suffixes appended}

\item{write.tsv}{logical, whether to emit output phenotype data in tsv tab-delimited plaintext}

\item{write.stata}{logical, whether to emit output phenotype data in STATA .dta format}

\item{write.spss}{logical, whether to emit output phenotype data in SPSS .zsav format}

\item{write.sas}{logical, whether to emit output phenotype data in SAS .sas7bdat format,
along with a source .sas file that needs to be run to assign category levels and types}

\item{write.yaml}{logical, whether to emit final version of stored configuration data
in YAML format; currently not tested}
}
\description{
This function contains handlers for the primary supported output formats
for this package: plaintext tab-delimited, STAT (.dta), SAS (.xpt only),
SPSS (TBD), and YAML (for final configuration; TBD). Additional control
parameters are likely to be added at a later data. Any number of outputs
can be selected and they will all be reported sequentially.
}
